# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for pack sizes and order pack calculation.",
        "title": "RE Partners API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/orders/calculate": {
            "post": {
                "description": "Returns minimum total items and pack breakdown for requested order size.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Calculate order packs",
                "parameters": [
                    {
                        "description": "Order calculation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.Calculate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CalculateResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/packs": {
            "get": {
                "description": "Returns pack sizes ordered ascending.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "packs"
                ],
                "summary": "List pack sizes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.PackSize"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a pack size used for order calculations.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "packs"
                ],
                "summary": "Add pack size",
                "parameters": [
                    {
                        "description": "Pack size request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddPackSize"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/packs/{id}": {
            "delete": {
                "description": "Deletes pack size by identifier.",
                "tags": [
                    "packs"
                ],
                "summary": "Delete pack size",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pack size id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AddPackSize": {
            "type": "object",
            "properties": {
                "size": {
                    "type": "integer"
                }
            }
        },
        "dto.Calculate": {
            "type": "object",
            "properties": {
                "items_ordered": {
                    "type": "integer"
                }
            }
        },
        "dto.CalculateResult": {
            "type": "object",
            "properties": {
                "items_ordered": {
                    "type": "integer"
                },
                "packs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Pack"
                    }
                },
                "total_items": {
                    "type": "integer"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
        "dto.Pack": {
            "type": "object",
            "properties": {
                "quantity": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "dto.PackSize": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                }
            }
        }
    }
}